==> 1_download_data_with_selenium.ipynb <==
from selenium import webdriver
from selenium.webdriver.common.keys import Keys
import pandas as pd
import numpy as np
import datetime
import urllib
import time

# Open browser
driver = webdriver.Chrome('chromedriver.exe')

# Lets create a calendar of dates with 2 weeks frequency
cal = pd.date_range(start='2015-01-01', end='2020-01-01', freq='14D')

# Loop over the website
for i in range(len(cal)-1):
    
    # Open the website
    driver.get('http://stat.gibdd.ru')
    
    # Find and move to the right category of the website
    section = driver.find_element_by_id('downloadAction')
    section.click()

    # Find and move to the subcategory
    subcat = driver.find_element_by_xpath('/html/body/div[4]/div[3]/div/div[4]/div/a[3]')
    subcat.click()

    # Define date strings
    send_date_1 = cal[i].strftime('%d') + cal[i].strftime('%m') + cal[i].strftime('%Y')
    send_date_2 = cal[i+1].strftime('%d') + cal[i+1].strftime('%m') + cal[i+1].strftime('%Y')

    # Find input form to pass dates
    date_1 = driver.find_element_by_xpath('/html/body/div[4]/div[3]/div/div[4]/div[3]/div[2]/div/input[1]')
    date_2 = driver.find_element_by_xpath('/html/body/div[4]/div[3]/div/div[4]/div[3]/div[2]/div/input[2]')

    # Set date range
    date_1.click()
    date_1.send_keys(send_date_1)
    date_1.send_keys(Keys.ENTER)
    date_2.click()
    date_2.send_keys(send_date_2)
    date_2.send_keys(Keys.ENTER)

    # Set region
    search = driver.find_element_by_xpath('/html/body/div[4]/div[3]/div/div[4]/div[4]/div[2]/div/div[1]/input')
    search.send_keys('Астраханская область')
    region = driver.find_element_by_xpath('/html/body/div[4]/div[3]/div/div[4]/div[4]/div[2]/div/div[2]/div/div/div/ul/li/ul/li[3]/ul/li[5]/span')
    region.click()

    # Set type of data to download
    search = driver.find_element_by_xpath('/html/body/div[4]/div[3]/div/div[4]/div[8]/div[2]/div/div[1]/input')
    search.send_keys('Дорожно-транспортные происшествия')
    dtp = driver.find_element_by_xpath('/html/body/div[4]/div[3]/div/div[4]/div[8]/div[2]/div/div[2]/div/div/div/ul/li[1]/span[2]')
    dtp.click()

    # Choose the type of data structure
    button = driver.find_element_by_xpath('/html/body/div[4]/div[3]/div/div[4]/div[9]/div[2]/div/div/button[2]')
    button.click()

    # Choose the format of file
    xlsx = driver.find_element_by_xpath('/html/body/div[4]/div[3]/div/div[4]/div[10]/div[2]/div/div/button[2]')
    xlsx.click()
    
    # Click download button
    download = driver.find_element_by_xpath('/html/body/div[4]/div[3]/div/div[4]/div[11]/button')
    download.click()
    
    # Give some time to download the file
    time.sleep(10)
